bug's fix

+++ Services/FileService.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.Serialization;
 using System.Text;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
@@ -18,14 +19,18 @@ namespace LoggerApp.Services
             SerializationSample();
             DirectoryPath();
             DirectoryCheck();
-            FileCreate();
+            CreateFiles();
+            DeleteFiles(LogDirPath);
+            DeleteFiles(BackUpDirPath);
         }
+
         private static void SerializationSample()
         {
             var path = "Services\\jsconfig1.json";
             var configFile = File.ReadAllText(path);
             Config = JsonConvert.DeserializeObject<Config>(configFile);
         }
+
         private static void DirectoryPath()
         {
             LogDirPath = Config.Logger.DirectoryPath;
@@ -34,6 +39,7 @@ namespace LoggerApp.Services
             LogDirPath = LogDirPath.Trim('/');
             BackUpDirPath = BackUpDirPath.Trim('/');
         }
+
         private static void DirectoryCheck()
         {
             if (!Directory.Exists(LogDirPath))
@@ -41,39 +47,47 @@ namespace LoggerApp.Services
                 Console.WriteLine("Enter");
                 Directory.CreateDirectory(LogDirPath);
             }
+
             if (!Directory.Exists(BackUpDirPath))
             {
                 Console.WriteLine("Enter");
                 Directory.CreateDirectory(BackUpDirPath);
             }
         }
-        private static void FileCreate()
+
+        private static void CreateFiles()
         {
+            FileName filename = new FileName();
+            string name = filename.SetFromConfig(Config);
+
             var textLog = Logger.Sb.ToString();
 
-            var fullDate = DateTime.Now;
-            string timeString = fullDate.ToLongTimeString();
-            string dateString = fullDate.ToShortDateString();
-            timeString = timeString.Replace(':', '.');
-
-            string fileName = $"{timeString} {dateString}";
-
-            File.WriteAllText($"{LogDirPath}\\{fileName}.txt", textLog);
-            File.WriteAllText($"{BackUpDirPath}\\{fileName}.txt", textLog);
-
-            FilesCounter(LogDirPath);
-            FilesCounter(BackUpDirPath);
+            File.WriteAllText($"{LogDirPath}\\{name}.txt", textLog);
+            File.WriteAllText($"{BackUpDirPath}\\{name}.txt", textLog);
         }
-        private static void FilesCounter(string dirName)
+
+        private static void DeleteFiles(string dirName)
         {
             string[] filesArray = Directory.GetFiles(dirName);
 
             if (filesArray.Length > 3)
-            {   
+            {
+                DateTime[] dateArr = new DateTime[filesArray.Length];
+                for (int i = 0; i < filesArray.Length; i++)
+                {
+                    dateArr[i] = File.GetCreationTime(filesArray[i]);
+                }
+
                 int counterToDelFiles = filesArray.Length - 3;
                 for (int i = 0; i < counterToDelFiles; i++)
                 {
-                    File.Delete(filesArray[i]);
+                    for (int j = 0; j < filesArray.Length; j++)
+                    {
+                        if (dateArr[i].CompareTo(dateArr[j]) < 0)
+                        {
+                            File.Delete(filesArray[i]);
+                        }
+                    }
                 }
             }
         }
